Create Ele.js
I often make a lot of elements in my scripts. I wrote this function to cut down on all of the verbose text involved in doing this. 
This function takes a tag name and a json object as parameters and returns a shiny new element. I find it much easier to use than the traditional method.